Refactor package structure for ApiConstants; update imports in related classes
