+++ infrastructure/entry-points/reactive-web/src/main/java/co/com/pragma/api/constants/ApiConstants.java
@@ -1,4 +1,4 @@
-package co.com.pragma.api;
+package co.com.pragma.api.constants;
 
 public class ApiConstants {
